update of chatApp

+++ base/views.py
@@ -1,7 +1,9 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
 from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from .models import Room, Topic
 from .forms import RoomForm
@@ -9,6 +11,9 @@ from .forms import RoomForm
 
 
 def loginPage(request):
+    
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -16,9 +21,20 @@ def loginPage(request):
             user = User.objects.get(username=username)
         except:
             messages.error(request, 'User does not exist')
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Invalid username or password')
     context = {}
     return render(request, 'base/login_register.html', context)
 
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+    
 def home(request):
     
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -36,6 +52,7 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -48,9 +65,13 @@ def createRoom(request):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -61,9 +82,12 @@ def updateRoom(request, pk):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
     
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
